API/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
sqlalchemy>=2.0.28
pydantic>=2.6.1
pydantic-settings>=2.1.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.7
email-validator>=2.1.0
bcrypt>=4.1.2
alembic>=1.13.1
typing-extensions>=4.9.0
